Minor xmp  fixes (#1009)

+++ src/icloudpd/xmp_sidecar.py
@@ -37,7 +37,7 @@ class XMPMetadata(NamedTuple):
 
 
 def generate_xmp_file(
-    logger: logging.Logger, download_path: str, asset_record: dict[str, Any]
+    logger: logging.Logger, download_path: str, asset_record: dict[str, Any], dry_run: bool
 ) -> None:
     sidecar_path: str = download_path + ".xmp"
     can_write_file: bool = True
@@ -56,21 +56,22 @@ def generate_xmp_file(
     # decode asset record fields
     # for k in asset_record['fields']:
     #    if asset_record["fields"][k]['type'] == "ENCRYPTED_BYTES":
-    # try:
-    #     asset_record["fields"][k]['decoded'] = plistlib.loads(base64.b64decode(asset_record['fields'][k]['value']), fmt=plistlib.FMT_BINARY)
-    # except plistlib.InvalidFileException:
-    #     try:
-    #         asset_record["fields"][k]['decoded'] =  json.loads(zlib.decompress(base64.b64decode(asset_record['fields'][k]['value']),-zlib.MAX_WBITS))
-    #     except Exception as e:
-    #         asset_record["fields"][k]['decoded'] = base64.b64decode(asset_record['fields'][k]['value']).decode("utf-8")
+    #         try:
+    #             asset_record["fields"][k]['decoded'] = plistlib.loads(base64.b64decode(asset_record['fields'][k]['value']), fmt=plistlib.FMT_BINARY)
+    #         except plistlib.InvalidFileException:
+    #             try:
+    #                 asset_record["fields"][k]['decoded'] =  json.loads(zlib.decompress(base64.b64decode(asset_record['fields'][k]['value']),-zlib.MAX_WBITS))
+    #             except Exception as e:
+    #                 asset_record["fields"][k]['decoded'] = base64.b64decode(asset_record['fields'][k]['value']).decode("utf-8")
     # json.dump(asset_record["fields"],         open(download_path + ".ar.json", "w"),         indent=4,        default=str,        sort_keys=True)
 
     if can_write_file:
         xmp_metadata: XMPMetadata = build_metadata(asset_record)
         xml_doc: ElementTree.Element = generate_xml(xmp_metadata)
-        # Write the XML to the file
-        with open(sidecar_path, "wb") as f:
-            f.write(ElementTree.tostring(xml_doc, encoding="utf-8", xml_declaration=True))
+        if not dry_run:
+            # Write the XML to the file
+            with open(sidecar_path, "wb") as f:
+                f.write(ElementTree.tostring(xml_doc, encoding="utf-8", xml_declaration=True))
 
 
 def build_metadata(asset_record: dict[str, Any]) -> XMPMetadata:
@@ -86,8 +87,15 @@ def build_metadata(asset_record: dict[str, Any]) -> XMPMetadata:
             "utf-8"
         )
 
+    # adjustementSimpleDataEnc can be one of three formats:
+    # - a binary plist - starting with 'bplist00' ( YnBsaXN0MD once encoded), seemingly used for some videos metadata (slow motion range etc)
+    # - a CRDT (Conflict-free Replicated Data Types) - starting with 'crdt' (Y3JkdA once encoded) - used for drawings and annotations on photos and screenshots
+    # - a zlib compressed JSON - used for simple photo metadata adjustments (orientation etc) 
+    # for exporting metadata, we only consider the JSON data, but it's the only one that doesn't have a predictable start pattern, so we check by excluding the other two
     orientation = None
-    if "adjustmentSimpleDataEnc" in asset_record["fields"]:
+    if ("adjustmentSimpleDataEnc" in asset_record["fields"] 
+        and not asset_record["fields"]["adjustmentSimpleDataEnc"]["value"].startswith("Y3JkdA") # "crdt"
+        and not asset_record["fields"]["adjustmentSimpleDataEnc"]["value"].startswith("YnBsaXN0MD")): # "bplist00"
         adjustments = json.loads(
             zlib.decompress(
                 base64.b64decode(asset_record["fields"]["adjustmentSimpleDataEnc"]["value"]),
